Solved bug price

+++ SSD_POLI_API/DailyMenuModel.cs
@@ -3,8 +3,8 @@
     public int Id { get; set; }
     public string? Title { get; set; }
     public string Description { get; set; }
-    public decimal PriceForUPT { get; set; }
-    public decimal PriceOutsidersUPT { get; set; }
+    public double PriceForUPT { get; set; }
+    public double PriceOutsidersUPT { get; set; }
     public byte[]? Picture { get; set; }
 
 }
