+++ SSD_POLI_API/StandardMenuModel.cs
@@ -5,8 +5,8 @@ public class StandardMenuModel
     public int Id { get; set; }
     public string Title { get; set; }
     public string Description { get; set; }
-    public decimal PriceForUPT { get; set; }
-    public decimal PriceOutsidersUPT { get; set; }
+    public double PriceForUPT { get; set; }
+    public double PriceOutsidersUPT { get; set; }
     public byte[]? Picture { get; set; }
 
 }
